Check email block list when email sign-up

+++ plume-models/src/email_signups.rs
@@ -1,4 +1,5 @@
 use crate::{
+    blocklisted_emails::BlocklistedEmail,
     db_conn::DbConn,
     schema::email_signups,
     users::{NewUser, Role, User},
@@ -60,6 +61,10 @@ pub struct NewEmailSignup<'a> {
 
 impl EmailSignup {
     pub fn start(conn: &DbConn, email: &str) -> Result<Token> {
+        if let Some(x) = BlocklistedEmail::matches_blocklist(conn, email)? {
+            return Err(Error::Blocklisted(x.notify_user, x.notification_text));
+        }
+
         conn.transaction(|| {
             Self::ensure_user_not_exist_by_email(conn, email)?;
             let _rows = Self::delete_existings_by_email(conn, email)?;
@@ -90,6 +95,9 @@ impl EmailSignup {
     }
 
     pub fn confirm(&self, conn: &DbConn) -> Result<()> {
+        if let Some(x) = BlocklistedEmail::matches_blocklist(conn, &self.email)? {
+            return Err(Error::Blocklisted(x.notify_user, x.notification_text));
+        }
         conn.transaction(|| {
             Self::ensure_user_not_exist_by_email(conn, &self.email)?;
             if self.expired() {
@@ -101,6 +109,9 @@ impl EmailSignup {
     }
 
     pub fn complete(&self, conn: &DbConn, username: String, password: String) -> Result<User> {
+        if let Some(x) = BlocklistedEmail::matches_blocklist(conn, &self.email)? {
+            return Err(Error::Blocklisted(x.notify_user, x.notification_text));
+        }
         conn.transaction(|| {
             Self::ensure_user_not_exist_by_email(conn, &self.email)?;
             let user = NewUser::new_local(
